Split into NcMath and NcUtility (#4)
* need public build interface

* update directories

* update readme
