+++ include/ncmath/DirectXMathUtility.h
