+++ include/ncmath/Geometry.h
