+++ include/ncmath/Quaternion.h
