+++ include/ncutility/Hash.h
