+++ include/ncutility/platform/SourceLocation.h
