--- source/math/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(NcCommon
-    PRIVATE
-        Quaternion.cpp
-)
